Refactor AppUtilMixin to shared file and fix its usage (#389)
Mixins should be listed in class definition lists prior to the
superclass. Since other components may also need mixins,
they have been moved to a new file.
